Split the major components into different files.
This is very much not complete, but seems to be functioning as well as
it was before, so this is a good starting point.
