+++ settings.py
@@ -1,0 +1,3 @@
+import os
+ADMIN_PASSWORD = os.environ['ADMIN_PASSWORD']
+ANNOTATOR_ID = 'annotator_id'
